Fix checking size

+++ main.go
@@ -1562,7 +1562,7 @@ func C_GetInterfaceList(pInterfaceList C.CK_INTERFACE_PTR, pulCount C.CK_ULONG_P
 		return C.CKR_ARGUMENTS_BAD
 	}
 
-	var INTERFACE_COUNT = len(interfaces)
+	var INTERFACE_COUNT = uint(len(interfaces))
 
 	if pInterfaceList == nil {
 		*pulCount = (C.CK_ULONG)(INTERFACE_COUNT)
@@ -1572,7 +1572,7 @@ func C_GetInterfaceList(pInterfaceList C.CK_INTERFACE_PTR, pulCount C.CK_ULONG_P
 	const CK_INTERFACE_SIZE = C.sizeof_CK_INTERFACE
 
 	*pulCount = (C.CK_ULONG)(INTERFACE_COUNT)
-	if getSizeOf(*pInterfaceList) < INTERFACE_COUNT*int(CK_INTERFACE_SIZE) {
+	if getSizeOf(*pInterfaceList) < INTERFACE_COUNT*uint(CK_INTERFACE_SIZE) {
 		return C.CKR_BUFFER_TOO_SMALL
 	}
